third_party/blink/web_tests/http/tests/devtools/sources/debugger/live-edit-no-reveal-expected.txt

Tests live edit feature.


Running: testLiveEditWithoutStepInWhenPausedThenStepIntoCausesCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Script execution paused.
Stepping into...
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (2, 4).

Running: testLiveEditWithStepInWhenPausedThenStepIntoCausesCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Did step into
Stepping into...
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (8, 4).

Running: testLiveEditWithoutStepInWhenPausedDoesNotCauseCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Script execution paused.
Script execution resumed.
Cursor position is: (0, 0).

Running: testLiveEditWithStepInWhenPausedDoesNotCauseCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Did step into
Script execution resumed.
Script execution paused.
Script execution resumed.
Cursor position is: (0, 0).
